Major template change
This repo initially contained a cloned version of greek yoghurt to
which I was adding bootstrap.
In this version I have started a new theme from scratch.
To do this I modified one of Bootstrap standard templates and then
added Concrete 5 areas.
Still need to work on the main navigation menu and the banner area.